Вынес класс логирования в бд в отдельный файл
